requirements.txt
langgraph>=0.2.0
langchain-ollama>=0.2.0
langchain-core>=0.3.0
langsmith>=0.1.0
PyGithub>=2.1.0
requests>=2.31.0
pyyaml>=6.0
python-dotenv>=1.0.0
pytest>=8.0.0
notebook>=7.0.0
ipykernel>=6.0.0
graphviz>=0.20.0
